Move IsolateStageLights to CameraQueueComponent where it is used
