+++ RuntimeIcons/src/Components/CameraQueueComponent.cs
@@ -16,6 +16,7 @@ using RuntimeIcons.Utils;
 using UnityEngine;
 using UnityEngine.Experimental.Rendering;
 using UnityEngine.Rendering;
+using UnityEngine.Rendering.HighDefinition;
 
 namespace RuntimeIcons.Components;
 
@@ -340,7 +341,7 @@ public class CameraQueueComponent : MonoBehaviour
         PrepareNextRender();
     }
 
-    private StageComponent.IsolateStageLights _isolatorHolder;
+    private IsolateStageLights _isolatorHolder;
 
     private void OnBeginCameraRendering(ScriptableRenderContext context, Camera camera)
     {
@@ -377,7 +378,7 @@ public class CameraQueueComponent : MonoBehaviour
             // Set the item on the stage and isolate lighting.
             Stage.SetStageFromSettings(_nextRender.Value.Settings);
 
-            _isolatorHolder = new StageComponent.IsolateStageLights(settings.TargetObject.gameObject, Stage.LightGo);
+            _isolatorHolder = new IsolateStageLights(settings.TargetObject.gameObject, Stage.LightGo);
             _isStaged = true;
         }
         catch (Exception ex)
@@ -651,6 +652,51 @@ public class CameraQueueComponent : MonoBehaviour
                 using var markerAuto = TryDequeueReadyMarker.Auto();
             #endif
             return ReadyQueue.TryDequeue(out settings);
+        }
+    }
+
+    internal class IsolateStageLights : IDisposable
+    {
+        private List<Light> _lightMemory;
+        private Color _ambientLight;
+
+        public IsolateStageLights(params GameObject[] stageObjects)
+        {
+            _lightMemory = UnityEngine.Pool.ListPool<Light>.Get();
+
+            _ambientLight = RenderSettings.ambientLight;
+            RenderSettings.ambientLight = Color.black;
+
+            var localLights = UnityEngine.Pool.ListPool<Light>.Get();
+            foreach (var stageObject in stageObjects)
+                localLights.AddRange(stageObject.GetComponentsInChildren<Light>());
+
+            foreach (var light in FindObjectsByType<Light>(FindObjectsSortMode.None))
+            {
+                if (!light.enabled)
+                    continue;
+                if (localLights.Contains(light))
+                    continue;
+                _lightMemory.Add(light);
+                light.enabled = false;
+            }
+
+            UnityEngine.Pool.ListPool<Light>.Release(localLights);
+        }
+
+        public void Dispose()
+        {
+            if (_lightMemory == null)
+                return;
+
+            RenderSettings.ambientLight = _ambientLight;
+            _ambientLight = Color.black;
+
+            foreach (var light in _lightMemory)
+                light.enabled = true;
+
+            ListPool<Light>.Release(_lightMemory);
+            _lightMemory = null;
         }
     }
